refactor(composables): extract reusable logic from MainSideBar and Settings
- Create useNavigationLinks composable for shared navigation links
- Create useSidebarResize composable for sidebar drag-to-resize
- Create useSaveListExpandBehavior composable for menu expand state
- Refactor MainSideBar.vue to use new composables (~140 lines reduced)
- Refactor Settings.vue to use useNavigationLinks
